:recycle: RacingCarGameController 분리
